+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Current
 * Have the server ensure it only sends exact chunk-sized key data.
 * Add further checks on stream parameters before serializing.
 * Handle messages similar to events, adding a "none" variant.
+* Add a field for user data to the stream and buffer structures.
 
 Version 0.3.1
 -------------